3_kms/requirements.txt
google-cloud-pubsub
canonicaljson
PyJWT
google-api-python-client
httplib2
oauth2client
requests
google-auth-httplib2
pycrypto
canonicaljson
